Refactor config.py into two modules
